Refactoring and file code splitting. Locks+voice change detection still finicky. Need more granular locking/unlocking
